requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
docopt==0.6.2
exceptiongroup==1.1.0
flake8==6.0.0
generators==2020.4.27
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
more-itertools==9.0.0
outcome==1.2.0
packaging==21.3
pipreqs==0.4.11
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
requests==2.28.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
strict-functions==2020.2.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
wsproto==1.2.0
yarg==0.1.9
